# public/openapi.yaml
openapi: 3.1.0
info:
  title: Q-SLICE Sovereign Cockpit API
  version: 2.2.0
  description: |
    Sovereign interface for CCCE metrics, language-locked agent wrappers,
    evidence ledger, and quantum threat assessment.

    SPEC_LOCK v2.2.0 | CAGE: 9HUP5 | Agile Defense Systems, LLC

    Physical Constants (Immutable):
      ΛΦ = 2.176435×10⁻⁸ s⁻¹
      Φ_threshold = 0.7734
      Γ_critical = 0.300
      θ_lock = 51.843°
servers:
  - url: https://q-slice-redteam-arena.vercel.app
    description: Q-SLICE Production Server
paths:
...
  /api/qslice:
    get:
      operationId: getQSliceData
      summary: Query Q-SLICE threat lab or evidence chain
      parameters:
        - name: action
          in: query
          schema:
            type: string
            enum: [backends, vectors, experiment, list, circuits, evidence, health]
      responses:
        '200':
          description: Q-SLICE status or data
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/EvidenceLedger'
                  - $ref: '#/components/schemas/QsliceHealth'

    post:
      operationId: submitQSliceExperiment
      summary: Submit a threat experiment (fail-closed in QPU mode)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - vector
              properties:
                vector:
                  type: string
                  enum: [QUANTUM_EXPLOITATION, SUBVERSION_OF_TRUST, INTEGRITY_DISRUPTION, COHERENCE_ATTACKS, ENTANGLEMENT_HIJACKING]
                shots:
                  type: integer
                  default: 1024
                backend:
                  type: string
                  enum: [ibm_fez, ibm_torino, ibm_marrakesh]
                simulate:
                  type: boolean
                  default: false
      responses:
        '200':
          description: Experiment submitted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  experiment_id:
                    type: string
                  evidence_id:
                    type: string
                  request_hash:
                    type: string
                  evidence_class:
                    type: string
                    enum: [CLASS_A, CLASS_B, CLASS_C]
        '403':
          description: Fail-closed policy violation
        '503':
          description: Environment check failed
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EvidenceEntry:
      type: object
      properties:
        id:
          type: string
        timestamp:
          type: string
          format: date-time
        class:
          type: string
          enum: [CLASS_A, CLASS_B, CLASS_C]
        request_hash:
          type: string
        result_hash:
          type: string
        leaf_hash:
          type: string
        chain_index:
          type: integer
    EvidenceLedger:
      type: object
      properties:
        chain_state:
          type: object
          properties:
            root:
              type: string
            leaf_count:
              type: integer
            spec_lock_hash:
              type: string
        evidence_count:
          type: integer
        recent_entries:
          type: array
          items:
            $ref: '#/components/schemas/EvidenceEntry'
    QsliceHealth:
      type: object
      properties:
        success:
          type: boolean
        mode:
          type: string
          enum: [QPU, DEV]
        fail_closed_status:
          type: object
          properties:
            allowed:
              type: boolean
            reason:
              type: string
        ibm_configured:
          type: boolean
        spec_lock:
          type: object